done with lab 2

+++ lab2/DebugExercise/DebugExercise1.java
@@ -6,11 +6,10 @@ package DebugExercise;
  */
 public class DebugExercise1 {
     public static int divideThenRound(int top, int bottom) {
-        int quotient = top / bottom;
+        float quotient = (float) top / bottom;
         int result = Math.round(quotient);
         return result;
     }
-
     public static void main(String[] args) {
         int t = 10;
         int b = 2;
